Updating comments for deprecations.
The `@todo` tag is used for internal messages for developers maintaining the library and the class, explaining that (and why) it should be removed in a future version. The `@deprecated` tag is intended for the eyes of someone using this library to tell them that the method is deprecated and giving an alternative suggestion.